refactor(serializers/user): user module
